refactor: move banner to /about dir
