+++ components/scoped/about/Banner/index.tsx
